Fix StyleCop warnings in Japanese\TimePeriodParser
